osmApi: refactor – move to `services/osm` folder
