+++ src/components/FeaturePanel/EditDialog/useGetHandleSave.tsx
@@ -3,8 +3,8 @@ import { useEditDialogFeature } from './utils';
 import { useEditContext } from './EditContext';
 import { createNoteText } from './createNoteText';
 import { t } from '../../../services/intl';
-import { saveChanges } from '../../../services/osmApiAuth';
-import { insertOsmNote } from '../../../services/osmApi';
+import { saveChanges } from '../../../services/osm/osmApiAuth';
+import { insertOsmNote } from '../../../services/osm/osmApi';
 import { useSnackbar } from '../../utils/SnackbarContext';
 import { getShortId } from '../../../services/helpers';
