+++ src/services/osm/__tests__/osmToFeature.test.ts
@@ -1,6 +1,6 @@
 import { osmToFeature } from '../osmToFeature';
 
-import { ImageDef } from '../types';
+import { ImageDef } from '../../types';
 
 const nodeResponse = {
   type: 'node',
